refactor(update): inline action.ts into index.tsx
Collapse the update command into a single file per review feedback; the logic
stays unit-testable via the injected ProcessRunner and a fetch spy. Tests import
from ./index.
